+++ frontend/src/index.css
@@ -10,8 +10,8 @@ body {
 }
 
 .modal-body {
-  columns: 2 auto;
-  column-gap: 10%;
+  max-height: calc(100vh - 210px);
+  overflow-y: auto;
 }
 
 .hide-scrollbar::-webkit-scrollbar {
